Moved deprecated featurespec tests to featurespec-test.
