+++ jvm/featurespec-test/src/test/scala/org/scalatest/featurespec/DeprecatedAsyncFeatureSpecLikeSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.featurespec
 
 import scala.concurrent.{Promise, ExecutionContext, Future}
 import org.scalatest._
