+++ jvm/featurespec-test/src/test/scala/org/scalatest/featurespec/DeprecatedAsyncFeatureSpecSpec2.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.featurespec
 
 import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest._
